feat: Add generateBackgroundPizzas().
Instead anonimous function in the sroll event listener
and add orientationchange listener for it. This was done
to optimize the number of background pizzas (and
layers respectively) according to device resolution.

Also getElementById of 'movingPizzas1' was taked out
into the global scope. There is no need to get new collection
every scroll event.